Made the camera work more like you'd expect from a fps camera

+++ util.h
@@ -44,13 +44,13 @@ typedef struct quaternion_s {
 	double x, y, z, w;
 } quaternion_t;
 
-int quad_length(quaternion_t *q);
+double quad_length(quaternion_t *q);
 void quad_normalize(quaternion_t *q);
 void quad_mult(quaternion_t *res, quaternion_t *a, quaternion_t *b);
 void quad_conjugate(quaternion_t *q);
 void quad_rotate(quaternion_t* q, double angle, double axis[3]);
 void quad_applyrotation(quaternion_t *res, quaternion_t *rot);
-int quad_diff(quaternion_t *q1, quaternion_t *q2);
+double quad_diff(quaternion_t *q1, quaternion_t *q2);
 
 /* FPS independent Animation */
 typedef struct anim_task_s {
